Airflow/requirements.txt
apache-airflow-providers-common-sql>=1.20.0
psycopg2-binary==2.9.10
asyncpg>=0.30.0
apache-airflow-providers-postgres==6.0.0
pika==1.3.2
trino==0.332.0